Not sure how to do evaluate yet
My idea is to put the values in the SymList which will be constantly
updated during run.

+++ src/created/ParseTree/SabiSabi/ss_a2.java
@@ -158,9 +158,9 @@ public abstract class ss_a2 implements created.iNode
             {
                 switch(((arithmetic_2.arithmetic2) ar).evaluate())
                 {
-                    case "*" : return (float)o1 * (float)o2;
-                    case "/" : return (float)o1 / (float)o2;
-                    case "%" : return (float)o1 % (float)o2;
+                    case "*" : return (double)o1 * (double)o2;
+                    case "/" : return (double)o1 / (double)o2;
+                    case "%" : return (double)o1 % (double)o2;
                 }
             }
